+++ apps/frontend/src/components/ui/shareable-resource/PublicShareableDocumentList.test.tsx
@@ -1,4 +1,4 @@
-import { ServiceListDisplayModeEnum } from '@/components/service/components/header/ServiceListHeader';
+import { ServiceListDisplayMode } from '@/components/service/components/header/ServiceListHeader';
 import testRender from '@/utils/test/test-render';
 import { publicDocumentListItemFragment$data } from '@generated/publicDocumentListItemFragment.graphql';
 import { seoServiceInstanceFragment$data } from '@generated/seoServiceInstanceFragment.graphql';
@@ -66,7 +66,7 @@ describe('PublicShareableDocumentList', () => {
         documents={documents}
         serviceInstance={serviceInstance}
         baseUrl={BaseUrl}
-        displayMode={ServiceListDisplayModeEnum.Tab}
+        displayMode={ServiceListDisplayMode.Tab}
       />
     );
 
@@ -98,7 +98,7 @@ describe('PublicShareableDocumentList', () => {
         documents={documents}
         serviceInstance={serviceInstance}
         baseUrl={BaseUrl}
-        displayMode={ServiceListDisplayModeEnum.List}
+        displayMode={ServiceListDisplayMode.List}
       />
     );
 
@@ -116,7 +116,7 @@ describe('PublicShareableDocumentList', () => {
         documents={documents}
         serviceInstance={serviceInstance}
         baseUrl={BaseUrl}
-        displayMode={ServiceListDisplayModeEnum.List}
+        displayMode={ServiceListDisplayMode.List}
       />
     );
 
@@ -136,7 +136,7 @@ describe('PublicShareableDocumentList', () => {
         documents={documents}
         serviceInstance={serviceInstance}
         baseUrl={BaseUrl}
-        displayMode={ServiceListDisplayModeEnum.List}
+        displayMode={ServiceListDisplayMode.List}
       />
     );
     const firstRow = screen.getByText(FirstDocumentName).closest('tr');
