+++ apps/frontend/src/components/ui/shareable-resource/PublicShareableResourceList.test.tsx
@@ -1,4 +1,4 @@
-import { ServiceListDisplayModeEnum } from '@/components/service/components/header/ServiceListHeader';
+import { ServiceListDisplayMode } from '@/components/service/components/header/ServiceListHeader';
 import testRender from '@/utils/test/test-render';
 import { publicDocumentListItemFragment$data } from '@generated/publicDocumentListItemFragment.graphql';
 import { seoServiceInstanceFragment$data } from '@generated/seoServiceInstanceFragment.graphql';
@@ -24,7 +24,7 @@ describe('PublicShareableResourceList', () => {
         documents={[]}
         serviceInstance={serviceInstance as seoServiceInstanceFragment$data}
         baseUrl="https://xtm.local"
-        displayMode={ServiceListDisplayModeEnum.Tab}
+        displayMode={ServiceListDisplayMode.Tab}
       />
     );
 
@@ -54,7 +54,7 @@ describe('PublicShareableResourceList', () => {
         documents={documents as publicDocumentListItemFragment$data[]}
         serviceInstance={serviceInstance as seoServiceInstanceFragment$data}
         baseUrl="https://xtm.local"
-        displayMode={ServiceListDisplayModeEnum.Tab}
+        displayMode={ServiceListDisplayMode.Tab}
       />
     );
